requirements.txt
Django==3.0.5
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytest-django==3.9.0
pytest==5.4.1
pyyaml
uritemplate
# pillow>=4.0.0
